chore(code): code organization
