+++ src/components/Observer/index.ts
@@ -1,0 +1,2 @@
+export * from "./Observer.interface";
+export * from "./Subject.class";
