+++ src/components/Plushies/SpiffoSanta.plushie.ts
@@ -1,0 +1,9 @@
+import { Plushie } from "./Plushie.class";
+
+
+export class SpiffoSanta extends Plushie {
+	public subscribe() {
+		super.subscribe();
+		print("SpiffoSanta buff should be applied");
+	}
+}
