+++ src/components/Plushies/index.ts
@@ -1,0 +1,2 @@
+export * from "./Plushie.class";
+export * from "./SpiffoSanta.plushie";
